% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_climate}
\alias{data_climate}
\alias{monthly_climate_basel}
\alias{monthly_climate_giessen}
\alias{Basel_Temperature_ets_resid}
\alias{monthly_co2_wide}
\title{Climate Example / test dataset}
\format{
data.frame
}
\source{
\strong{Federal Office of Meteorology and Climatology MeteoSwiss}
\url{https://github.com/CSSEGISandData/COVID-19/tree/master/csse_covid_19_data/csse_covid_19_time_series}

\strong{DWD Archiv Monats- und Tageswerte}
\url{https://www.dwd.de/DE/leistungen/klimadatendeutschland/klarchivtagmonat.html}

\strong{NOAA ESRL} \url{https://www.esrl.noaa.gov/gmd/ccgg/trends/global.html}
}
\usage{
data(monthly_climate_basel)

data(monthly_climate_giessen)

data(Basel_Temperature_ets_resid)

data(monthly_co2_wide)
}
\description{
The climate datasets provide Temperature and Precipitation time series data
from Germany (DWD), Switzerland (meteoswiss) and atmospheric CO_2 concentration
time series data from Mauna Loa (esrl).
}
\details{
\code{monthly_climate_basel} provides monthly Temperature and Precipitation
values (long format) from 2016 until 2020 without any NAs,
\cr
\code{monthly_climate_giessen} same as for Basel but with some NA values.

\code{Basel_Temperature_ets_resid} provides ETS model data with fitted and
residual values.

\code{monthly_co2_wide} provides monthly CO2 time series data (wide format)
from Mauna Loa from 1958 to 2020.
}
\examples{
# Temperature and Precipitation data
head(monthly_climate_basel)
tail(monthly_climate_basel)
uts_gen_yearly_seasonal_avg(monthly_climate_basel)

# Precipitation data have NAs
dplyr::slice(monthly_climate_giessen \%>\%
  dplyr::filter(Measure == "Precipitation"), 13:32)
uts_gen_yearly_seasonal_avg(monthly_climate_giessen)

# ETS model data
head(Basel_Temperature_ets_resid)
str(Basel_Temperature_ets_resid)
ggts_histo_forecast_resid(Basel_Temperature_ets_resid)

# CO_2 data wide format
monthly_co2_wide
}
\keyword{datasets}
