% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggts_corona.R
\name{ggts_conf_deaths_facet}
\alias{ggts_conf_deaths_facet}
\title{Trend facet plot for each case type and each country}
\usage{
ggts_conf_deaths_facet(
  data,
  x = .data$Date,
  y = .data$Cases,
  vars_1 = .data$Case_Type,
  vars_2 = .data$Country
)
}
\arguments{
\item{data}{A data frame}

\item{x}{Unquoted df data-variable to specify the
time index variable (\code{default: Date}).}

\item{y}{Unquoted df data-variable of the cases}

\item{vars_1}{Unquoted df data-variable of the case types}

\item{vars_2}{Unquoted df data-variable of the countries}
}
\value{
plot object of mode "\code{plot}"
}
\description{
grid plot Confirmed / Death for each country
}
\examples{
# Corona data of "Germany", "Italy", "United States of America")
ggts_conf_deaths_facet(corona_data_sel)
}
\seealso{
\link{ggts_trend_facet}
}
