% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggts_corona.R
\name{ggts_cum_daily}
\alias{ggts_cum_daily}
\title{Cumulative and daily data trend plot}
\usage{
ggts_cum_daily(
  data,
  y_cum = .data$Cases,
  y_daily = .data$Daily_Cases,
  mean_daily = .data$Mean_Daily_Cases,
  country,
  span = 7,
  weeks = 12,
  ...
)
}
\arguments{
\item{data}{A data frame}

\item{y_cum}{Unquoted \verb{column name} of the data frame's cumulative cases}

\item{y_daily}{Unquoted \verb{column name} of the data frame's daily cases}

\item{mean_daily}{Unquoted \verb{column name} of the data frame's daily rolling mean data}

\item{country}{Unquoted \verb{column name} of the data frame's countries}

\item{span}{Numeric, span used for rolling mean calculation}

\item{weeks}{Numeric, number of time range weeks weeks for the daily data,
dates are provided in column \code{default = Date}}

\item{...}{Variable pass through to \code{\link[=ggts_trend_facet]{ggts_trend_facet()}}.
Unquoted \verb{column name} of the data frame's dates.}
}
\value{
plot object of mode "\code{plot}"
}
\description{
Provide \strong{trend plot} for \emph{cumulative} and \emph{daily} cases with
facets of \code{vars_1 = Case_Type}
}
\examples{
# Corona data of "Germany")
ggts_cum_daily(corona_data, country = "Germany", weeks = 6)
}
\seealso{
\link{ggts_trend_facet}, \link{ggts_conf_deaths_facet} and \link[ggplot2:ggplot]{ggplot2::ggplot}
}
