% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggts_climate.R
\name{ggts_decomp}
\alias{ggts_decomp}
\title{Generate Time Series Decompression ggplot facet}
\usage{
ggts_decomp(data, ...)
}
\arguments{
\item{data}{stlplus() output list.}

\item{...}{Other arguments.}
}
\value{
plot object of mode `plot
}
\description{
Generates from stlplus() output list components $Data and $time
a ggplot facet_wrap output over time for
"raw" (=original data), "seasonal", "trend" and "remainder" components.
}
\examples{
require(stlplus)
require(fpp3)
economics_tsbl <- ggplot2::economics \%>\%
 mutate(Year_Month = yearmonth(date)) \%>\%
 as_tsibble(index = Year_Month)
economics_tsbl \%>\% has_gaps() # check for time series gaps, fill with fill_gaps()
start_ts <- c(year(min(as.Date(economics_tsbl$Year_Month))),
  month(min(as.Date(economics_tsbl$Year_Month))))

# convert to time-series object
economics_ts <- ts(economics_tsbl$uempmed, start = start_ts, frequency = 12)
# convert to tibble object
economics_stlplus <- stlplus(economics_ts, s.window = 27, t.window = 1201)
#
ggts_decomp(uts_stlplus_as_tibble(economics_stlplus))

economics_stlplus <- stlplus(economics_ts, s.window = "per", t.window = 1201)
ggts_decomp(uts_stlplus_as_tibble(economics_stlplus))
}
\seealso{
\link{uts_stlplus_as_tibble} and \link[stlplus:stlplus]{stlplus::stlplus}
}
\keyword{misc}
