% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggts_corona.R
\name{ggts_logscale}
\alias{ggts_logscale}
\title{plot countries on log10scale}
\usage{
ggts_logscale(
  data,
  x = .data$Date,
  y = .data$Cases,
  vars_1 = .data$Country,
  vars_2 = .data$Case_Type
)
}
\arguments{
\item{data}{A data frame}

\item{x}{Unquoted df data-variable to specify the
time index variable (\code{default: Date}).}

\item{y}{Unquoted df data-variable of the cases}

\item{vars_1}{Unquoted df data-variable of the case types}

\item{vars_2}{Unquoted df data-variable of the countries}
}
\value{
plot object of mode "\code{plot}"
}
\description{
plot countries on log10scale
}
\examples{
# Corona data of "Germany", "Italy", "United States of America")
ggts_logscale(corona_data_sel, y = Cases)
}
