% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggts_climate.R
\name{ggts_season}
\alias{ggts_season}
\title{Provides Season (Winter, Spring, Summer, Fall) line and smooth plot
in one commmon diagram}
\usage{
ggts_season(data)
}
\arguments{
\item{data}{data frame with yearly data (long format) with average values
for the seasons.}
}
\description{
Provides Season (Winter, Spring, Summer, Fall) line and smooth plot
in one commmon diagram
}
\examples{
ggts_season(uts_gen_yearly_seasonal_avg(monthly_climate_basel))  +
  ggplot2::facet_wrap(~Measure, ncol = 1, scales = "free")
}
