% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggts_corona.R
\name{ggts_trend_facet}
\alias{ggts_trend_facet}
\title{Trend facet plot for each case type}
\usage{
ggts_trend_facet(
  data,
  x = .data$Date,
  y = .data$Cases,
  vars_1 = .data$Case_Type
)
}
\arguments{
\item{data}{A data frame}

\item{x}{Unquoted \verb{column name} of the data frame's dates}

\item{y}{Unquoted \verb{column name} of the data frame's cases}

\item{vars_1}{Unquoted \verb{column name} of the data frame's case types}
}
\value{
plot object of mode "\code{plot}"
}
\description{
Provide trend facet plot for each case type (e.g. Confirmed and Deaths)
}
\examples{
# Corona data of "Germany" - plot cumulative cases for 'Confirmed' and
# 'Deaths' and 'Confirmed'
ggts_trend_facet(corona_data \%>\% dplyr::filter(Country == "Germany"))
ggts_trend_facet(corona_data \%>\% dplyr::filter(Country == "Germany"), y = Cases_100k)
}
\seealso{
\link{ggts_cum_daily}
}
