% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggts_climate.R
\name{ggts_year_over_month}
\alias{ggts_year_over_month}
\title{Year over Month plot}
\usage{
ggts_year_over_month(data, time = Month, y = count, period = Period)
}
\arguments{
\item{data}{data frame with yearly data (long format) with average values
for the seasons.}

\item{time}{data frame column for values of x axes.}

\item{y}{data frame column for values of y axes.}

\item{period}{group and col aestetic for points and lines}
}
\value{
plot object of mode `plot
}
\description{
Year over Month plot
}
\examples{
ggts_year_over_month(monthly_climate_basel \%>\%
  dplyr::filter(Measure == "Temperature"), period = Year)
}
