% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggts_corona.R
\name{hc_bar_chart_country}
\alias{hc_bar_chart_country}
\title{Bar Chart plot with highchart()}
\usage{
hc_bar_chart_country(data, x = "Country", y, title = "Text", n = 15)
}
\arguments{
\item{data}{A data.frame object with data to chart. Code region and value are required.}

\item{x}{A string value with the column name for the bar names,
default = "Country".}

\item{y}{A string value with the column name for the bar data values.}

\item{title}{A string value with the title of the chart.}

\item{n}{numeric, number of bars in chart}
}
\value{
highchart plot
}
\description{
Visualization with top x country series bar chart
}
\examples{
last_date <- max(corona_data$Date)
data <- corona_data \%>\%
  dplyr::filter(Date == last_date & Case_Type == "Confirmed" & Country != "World")
hc_bar_chart_country(data, y = "Population",  title = "World Population")
}
\seealso{
\strong{Create a highchart object}
\url{https://rdrr.io/cran/highcharter/man/hchart.html}

\strong{Highcharts} \emph{series.bar} and \emph{aesthetic mappings}
\url{https://api.highcharts.com/highcharts/series.bar}
\url{https://rdrr.io/cran/highcharter/man/hcaes.html}
}
