% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggts_corona.R
\name{hc_world_map_plot}
\alias{hc_world_map_plot}
\title{highcharter World Map plot}
\usage{
hc_world_map_plot(data, value, title = "Text")
}
\arguments{
\item{data}{tbd}

\item{value}{tbd}

\item{title}{tbd}
}
\value{
highchart plot
}
\description{
highcharter World Map plot
}
\examples{
# so far no example
}
\references{
\strong{Highcharts} \emph{CONFIGURATION OPTIONS:}
\url{https://api.highcharts.com/highcharts/title}

\strong{Manual pages for highcharter}
\url{https://rdrr.io/cran/highcharter/man/hc_xAxis.html}
}
