% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{newpkg_package}
\alias{newpkg_package}
\title{New package provides function for \strong{Corona} data analysis}
\description{
Package creates documentation with roxygen2.

\strong{Markdown} is running also without \verb{@md}\emph{-tag}.
}
\details{
Details are still not outlined.

For an overview of R procedures and testthat files start with the vignettes:
browseVignettes(package = "newpkg"). However, so far no vignette is written.
}
\seealso{
Different links w/ and w/o backticks:

Main plot functions based on \link[ggplot2:ggplot]{ggplot2::ggplot} for
\code{\link[ggplot2:ggplot]{ggplot2::ggplot}} time series data:

\link{ggts_cum_daily}, \code{\link[=ggts_trend_facet]{ggts_trend_facet()}} and \link{ggts_conf_deaths_facet}

\code{\link{ggts_cum_daily}}, \code{\link[=ggts_trend_facet]{ggts_trend_facet()}} and \code{\link{ggts_conf_deaths_facet}}

and utility functions for time series:

\link{uts_get_corona_data_wide}, \link{uts_repronum}
}
\author{
Wolfgang Vollmer \email{wo.vollmer@online.de}
}
\keyword{package}
