% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{import}
\name{pkg-imports}
\alias{pkg-imports}
\alias{\%>\%}
\title{Imports from other packages}
\description{
Objects imported from other packages
}
\details{
These objects are imported from other packages.
Follow the links to their documentation.
\itemize{
\item magrittr forward-pipe operator: \code{\link[magrittr:pipe]{magrittr::\%>\%}}
}
}
