% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pkg_data}
\alias{pkg_data}
\alias{corona_data}
\alias{corona_data_sel}
\title{Example / test dataset}
\format{
data.frame
}
\source{
Johns Hopkins University on GitHub
\url{https://github.com/CSSEGISandData/COVID-19/tree/master/csse_covid_19_data/csse_covid_19_time_series}
}
\usage{
data(corona_data)

data(corona_data_sel)
}
\description{
The datasets provide the example test data
generated from the Johns Hopkins University data on GitHub.
}
\details{
\code{corona_data} provides data from all countries,
\code{corona_data_sel} provides data from a few selected countries only

see \href{https://wovollmer.github.io/github.io/#countries---table-overview}{Wolfgang's Corona Dashboard}
}
\examples{
head(corona_data)
uts_get_corona_data_wide(corona_data \%>\% dplyr::filter(Country == "Germany"))
head(corona_data_sel)
unique(corona_data_sel$Country)
}
\keyword{datasets}
