% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggts_corona.R
\name{plot_dygraph_daily}
\alias{plot_dygraph_daily}
\title{dygraph plot for Daily Confirmed and Death Cases}
\usage{
plot_dygraph_daily(data_xts, country_select, last_date, span = 7, weeks = 12)
}
\arguments{
\item{data_xts}{tbd}

\item{country_select}{tbd}

\item{last_date}{tbd}

\item{span}{tbd}

\item{weeks}{tbd}
}
\value{
tbd
}
\description{
dygraph interactive plot for xts time series objects:
input must be a named list or data frame, where the first element/column
provides x-axis values and all subsequent elements/columns provide one or
more series of y-values.
}
\references{
\url{https://rstudio.github.io/dygraphs/gallery-series-highlighting.html}
}
