% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggts_corona.R
\name{plot_dygraph_daily_repro}
\alias{plot_dygraph_daily_repro}
\title{dygraph plot for Daily Confirmed Cases and Reproduction Number w/ CI}
\usage{
plot_dygraph_daily_repro(
  data_xts,
  country_select,
  last_date,
  span = 7,
  weeks = 12
)
}
\arguments{
\item{data_xts}{tbd}

\item{country_select}{tbd}

\item{last_date}{tbd}

\item{span}{tbd}

\item{weeks}{tbd}
}
\value{
tbd
}
\description{
dygraph plot for Daily Confirmed Cases and Reproduction Number w/ CI
}
\references{
\url{https://rstudio.github.io/dygraphs/gallery-series-highlighting.html}
}
