% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uts_climate.R
\name{uts_gen_yearly_seasonal_avg}
\alias{uts_gen_yearly_seasonal_avg}
\title{Utilities for climate analysis and data reforming}
\usage{
uts_gen_yearly_seasonal_avg(data, season = TRUE)
}
\arguments{
\item{data}{A data frame with monthly "count" in long format}

\item{season}{Logical. If TRUE, the default, seasonal mean will be calculated}
}
\value{
data frame and tibble (not a tsibble), e.g.: \cr
\verb{A tibble: 10 x 8}
\verb{Groups:   City, Measure [2]}  # here: Measure = Temperature and Precipitation
\verb{City  Measure Year   Winter_avg Spring_avg Summer_avg Fall_avg Year_avg} \cr
\verb{<chr> <fct>   <dbl>  <dbl>      <dbl>      <dbl>      <dbl>    <dbl>}
}
\description{
"Widens" monthly long form data form and calculates
\itemize{
\item yearly average counts for Jan:Dec and
\item seasonal average counts mean data/month
\itemize{
\item seasons are Dec-Feb (DJF), Mar-May (MAM), June-Aug (JJA), Sept-Nov (SON)
\item Winter: year refers to January (realized by lag()).
}
\item note for Precipitation: to be multiplied by 12 if mm/year is needed
}
}
\details{
The input data frame must have the column names

\verb{City, Measure, Year_Month, Year} and \code{count}

and can be also a tibble or tsibble.
}
\examples{
# Temperature and Precipitation data of "Basel" w/o NAs rsp. Giessen w/ NAs
uts_gen_yearly_seasonal_avg(monthly_climate_basel)
uts_gen_yearly_seasonal_avg(monthly_climate_giessen)  # w/ NAs
# note: NAs of Winter_avg does not impact Year_avg if this
# NA is caused by December value of previous year
}
