% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uts_corona.R
\name{uts_get_corona_data_wide}
\alias{uts_get_corona_data_wide}
\title{Utilities for corona data analysis and data}
\usage{
uts_get_corona_data_wide(data)
}
\arguments{
\item{data}{A data frame to pivot}
}
\value{
data frame
}
\description{
Provide data in wide form and adapt column names
}
\details{
Data frame columns required are checked
}
\examples{
# Corona data of "Germany"
uts_get_corona_data_wide(corona_data \%>\% dplyr::filter(Country == "Germany"))
}
