% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uts_climate.R
\name{uts_interpolate_na}
\alias{uts_interpolate_na}
\title{Replace NA by interpolation separate for each time series}
\usage{
uts_interpolate_na(data, freq)
}
\arguments{
\item{data}{A data frame (column count)}

\item{freq}{season lenght, e.g. freq = 12 (1 year) for monthly temperature data}
}
\value{
data frame

if no NAs exist in count column => return unchanged data frame

otherwise: \cr
\verb{A tibble: 120 x 9}\cr
\verb{City Measure Year_Month  Year Month count   Raw NA_replace}\cr
\verb{<chr> <fct> <mth> <dbl> <fct> <dbl> <dbl> <dbl>}\cr

data frame with columns
\itemize{
\item count (updated) = orig and interplolated counts (replaced NAs, if any)
\item Raw (added) = orig counts (w/o replaced NAs)
\item NA_replace (added) = NAs (if count value exists) and
interplolated counts (replaced NAs, if any)
\item for plots w/ & w/o replaced counts
}
}
\description{
If there are there any NAs in monthly data set
=> replace NA by interpolation separate for each time series
}
\examples{
dplyr::slice(uts_interpolate_na(monthly_climate_giessen, freq =12), 75:90)
}
\references{
\strong{Schlittgen, Rainer - De Gruyter}, \emph{Angewandte Zeitreihenanalyse mit R},
2015 ISBN 978-3-11-041398-4
}
