% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uts_climate.R
\name{uts_ldrec}
\alias{uts_ldrec}
\title{Levinson-Durbin recursion}
\usage{
uts_ldrec(a)
}
\arguments{
\item{a}{value}
}
\description{
Levinson-Durbin recursion for determing all coefficients a(i,j),
mat <- ldrec(a)\cr
1<=i<=j<=maxlag (=p)\cr
input  : a is (p+1,1)-vector of acf of a time series: acov(0),...,acov(p)
or  1,acor(1),..,acor(p) \cr
output : (p,p+2)-matrix with coefficients in lower triangular,
pacf in colum p+2 and Q(p) in colum p+1\cr
}
\references{
Source: Rainer Schlittgen (01.10.2014) tsutil.r ldrec()
}
