% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uts_corona.R
\name{uts_repronum}
\alias{uts_repronum}
\title{Reproduction number calculation}
\usage{
uts_repronum(
  new.cases,
  profile,
  window = 1,
  delay = 0,
  conf.level = 0.95,
  pad.zeros = FALSE,
  min.denominator = 5,
  min.numerator = 5
)
}
\arguments{
\item{new.cases}{Vector of daily cases data}

\item{profile}{tbd}

\item{window}{tbd}

\item{delay}{tbd}

\item{conf.level}{tbd}

\item{pad.zeros}{tbd}

\item{min.denominator}{tbd}

\item{min.numerator}{tbd}
}
\value{
data.frame with as many rows as new.cases\preformatted{repronum = reproduction number

repronum.se = reproduction number standard error

ci.lower = confidence interval lower limit

ci.upper = confidence interval upper limit
}
}
\description{
Source: TU Ilmenau - GitHub
}
\details{
Here are further details
\subsection{subsection details and so on}{

However, needs some text.
\subsection{sub-subsection details}{

Also su-subsection needs some text.
}

}
}
\references{
\strong{TU Ilmenau} \emph{GitHub code source:} \url{https://github.com/Stochastik-TU-Ilmenau/COVID-19/blob/gh-pages/estimator.r}
}
