% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uts_climate.R
\name{uts_stlplus_as_tibble}
\alias{uts_stlplus_as_tibble}
\title{Converts stlplus() output object to a tibble object}
\usage{
uts_stlplus_as_tibble(stlplus_data)
}
\arguments{
\item{stlplus_data}{object of class "stlplus"}
}
\value{
\verb{# A tibble: 4,344 x 11} \cr
\verb{Year_Month Year Month count Raw Interpolated Trend Seasonal Remainder Seasonal_Adjust} \cr
\verb{<mth> <dbl> <fct> <dbl> <dbl>  <dbl> <dbl>  <dbl> <dbl> <dbl>} \cr
\verb{  1   1659 Jan  1659 Jan   3     3   2.84  9.41 -6.57 0.159 9.57} \cr

Easy plot with pkgTS::ggts_decomp()
}
\description{
New tibble object has NA replacements (if any) by
interpolation of seasonal plus trend values.
}
\examples{
require(stlplus)
require(fpp3)
economics_tsbl <- ggplot2::economics \%>\%
 mutate(Year_Month = yearmonth(date)) \%>\%
 as_tsibble(index = Year_Month)
economics_tsbl \%>\% has_gaps() # check for time series gaps, fill with fill_gaps()
start_ts <- c(year(min(as.Date(economics_tsbl$Year_Month))),
  month(min(as.Date(economics_tsbl$Year_Month))))

# convert to time-series object
economics_ts <- stats::ts(economics_tsbl$uempmed,
  start = start_ts, frequency = 12)
# convert to tibble object
economics_stlplus <- stlplus::stlplus(economics_ts, s.window = 27, t.window = 1201)
#
uts_stlplus_as_tibble(economics_stlplus)
ggts_decomp(uts_stlplus_as_tibble(economics_stlplus))
}
\seealso{
\link[stlplus:stlplus]{stlplus::stlplus} and \link{ggts_decomp}.
}
