% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggts_corona.R
\name{ggts_conf_deaths_facet}
\alias{ggts_conf_deaths_facet}
\title{Trend facet plot for each case type and each country}
\usage{
ggts_conf_deaths_facet(
  data,
  x = Date,
  y = Cases,
  vars_1 = Case_Type,
  vars_2 = Country
)
}
\arguments{
\item{data}{A data frame}

\item{x}{Unquoted \verb{column name} of the data frame's dates}

\item{y}{Unquoted \verb{column name} of the data frame's cases}

\item{vars_1}{Unquoted \verb{column name} of the data frame's case types}

\item{vars_2}{Unquoted \verb{column name} of the data frame's countries}
}
\value{
plot object of mode "\code{plot}"
}
\description{
grid plot Confirmed / Death for each country
}
\examples{
# Corona data of "Germany", "Italy", "United States of America")
ggts_conf_deaths_facet(corona_data_sel)
}
\seealso{
\link{ggts_trend_facet}
}
