% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggts_corona.R
\name{hc_bar_chart_country}
\alias{hc_bar_chart_country}
\title{Bar Chart plot with highchart()}
\usage{
hc_bar_chart_country(data, y, title = "Text", n = 15)
}
\arguments{
\item{data}{tbd}

\item{y}{tbd}

\item{title}{tbd}

\item{n}{tbd}
}
\value{
highchart plot
}
\description{
Visualization with top x country bar chart
}
\examples{
# example to get ordering plus World => via bind_rows(), currently not used:
#  data <- bind_rows(
#    data \%>\% dplyr::filter(Country != "World") \%>\%
#      arrange(desc(Cases_100k)) \%>\% head(14),
#    data \%>\%
#      dplyr::filter(Country == "World"))
last_date <- max(corona_data$Date)
data <- corona_data \%>\%
  dplyr::filter(Date == last_date & Case_Type == "Confirmed" & Country != "World")
hc_bar_chart_country(data, y = "Population",  title = "World Population")
}
\references{
\strong{Highcharts} \emph{CONFIGURATION OPTIONS:}
\url{https://api.highcharts.com/highcharts/title}

\strong{Manual pages for highcharter}
\url{https://rdrr.io/cran/highcharter/man/hc_add_series_map.html}

\url{https://rdrr.io/cran/highcharter/man/hc_xAxis.html}
}
