% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggts_corona.R
\name{hc_world_map_plot}
\alias{hc_world_map_plot}
\title{highcharter World Map plot}
\usage{
hc_world_map_plot(data, value, title = "Text")
}
\arguments{
\item{data}{A data.frame object with data to chart. Code region and value are required.}

\item{value}{A string value with the name of the variable to chart.}

\item{title}{A string value with the title of the chart.}
}
\value{
highchart plot
}
\description{
highcharter World Map plot
}
\examples{
last_date <- max(corona_data$Date)
data <- corona_data \%>\%
  dplyr::filter(Date == last_date & Case_Type == "Confirmed" &
                Country != "World")

data(worldgeojson, package = "highcharter")
hc_world_map_plot(data, value = "Population",  title = "World Population")

}
\references{
\strong{Highcharts} \emph{CONFIGURATION OPTIONS:}
\url{https://api.highcharts.com/highcharts/title}

\strong{Manual pages for highcharter}
\url{https://rdrr.io/cran/highcharter/man/hc_add_series_map.html}

\url{https://rdrr.io/cran/highcharter/man/hc_xAxis.html}
}
